+++ SQLiteStudio3/guiSQLiteStudio/datagrid/sqlqueryitemdelegate.cpp
@@ -6,6 +6,7 @@
 #include "sqlqueryview.h"
 #include "uiconfig.h"
 #include "common/utils_sql.h"
+#include "fkcombobox.h"
 #include "schemaresolver.h"
 #include <QHeaderView>
 #include <QPainter>
@@ -112,7 +113,7 @@ void SqlQueryItemDelegate::setEditorDataForFk(QComboBox* cb, const QModelIndex& 
 
 void SqlQueryItemDelegate::setModelData(QWidget* editor, QAbstractItemModel* model, const QModelIndex& index) const
 {
-    QComboBox* cb = dynamic_cast<QComboBox*>(editor);
+    FkComboBox* cb = dynamic_cast<FkComboBox*>(editor);
     QLineEdit* le = dynamic_cast<QLineEdit*>(editor);
     if (cb) {
         setModelDataForFk(cb, model, index);
@@ -126,29 +127,11 @@ void SqlQueryItemDelegate::setModelData(QWidget* editor, QAbstractItemModel* mod
     queryModel->notifyItemEditionEnded(index);
 }
 
-void SqlQueryItemDelegate::setModelDataForFk(QComboBox* cb, QAbstractItemModel* model, const QModelIndex& index) const
+void SqlQueryItemDelegate::setModelDataForFk(FkComboBox* cb, QAbstractItemModel* model, const QModelIndex& index) const
 {
-    SqlQueryModel* cbModel = dynamic_cast<SqlQueryModel*>(cb->model());
-    if (cbModel->isExecutionInProgress() || !cbModel->isAllDataLoaded())
-        return;
-
-    int idx = cb->currentIndex();
-    QModelIndex cbCol0Index = cbModel->index(idx, 0);
-    QString cbText = cb->currentText();
-    bool customValue = false;
-
-    if (cb->currentIndex() > -1 && !cbModel->itemFromIndex(cbCol0Index))
-    {
-        // Inserted QStandardItem by QComboBox, meaning custom value (out of dropdown model)
-        // With Qt 5.15 (maybe earlier) QComboBox started inserting QStandardItems and setting them as currentIndex.
-        // Here we're extracting this inserted value and remembering this is the custom value.
-        cbText = cbModel->data(cbCol0Index).toString();
-        customValue = true;
-    }
-
-    // Regardless if its preselected value or custom value, we need to honor empty=null setting
-    if (CFG_UI.General.KeepNullWhenEmptyValue.get() && model->data(index, Qt::EditRole).isNull() && cbText.isEmpty())
-        return;
+    bool manualValue = false;
+    bool ok = false;
+    QVariant value = cb->getValue(&manualValue, &ok);
 
     SqlQueryModel* dataModel = dynamic_cast<SqlQueryModel*>(model);
     SqlQueryItem* theItem = dataModel->itemFromIndex(index);
@@ -158,33 +141,19 @@ void SqlQueryItemDelegate::setModelDataForFk(QComboBox* cb, QAbstractItemModel* 
     if (!theItem)
     {
         qCritical() << "Confirmed FK edition, but there is no SqlQueryItem for which this was triggered!" << index;
-        model->setData(index, cbText, Qt::EditRole);
+        model->setData(index, value, Qt::EditRole);
         return;
     }
 
     // Out of index? So it's custom value. Set it and it's done.
     // If we deal with custom value inserted as item, we also just set it and that's it.
-    if (idx < 0 || idx >= cbModel->rowCount() || customValue)
+    if (manualValue)
     {
-        theItem->setValue(cbText);
+        theItem->setValue(value);
         return;
     }
 
-    // Otherwise we will have at least 2 columns. 1st column is hidden and is meta-column holding 1/0 (1 for value matching current cell value)
-    QList<SqlQueryItem *> row = cbModel->getRow(idx);
-    if (row.size() < 2 || !row[1])
-    {
-        // This happens when inexisting value is confirmed with "Enter" key,
-        // and rowCount() is apparently incremented, but items not yet.
-        // Very likely this was addressed in recent Qt versions (5.15 or a bit earlier)
-        // which resulted in value insertion and the "customValue" flag above in this method.
-        qCritical() << "Confirmed FK edition, but there is no combo item in the row for index" << idx << ", the item is" << theItem << ", CB row count is" << cbModel->rowCount();
-        theItem->setValue(cbText);
-        return;
-    }
-
-    QVariant comboData = row[1]->getValue();
-    theItem->setValue(comboData);
+    theItem->setValue(value);
 }
 
 void SqlQueryItemDelegate::setModelDataForLineEdit(QLineEdit* editor, QAbstractItemModel* model, const QModelIndex& index) const
@@ -274,7 +243,7 @@ QString SqlQueryItemDelegate::getSqlForFkEditor(SqlQueryItem* item) const
     QString col;
     QString firstSrcCol;
     QStringList usedNames;
-    for (SqlQueryModelColumn::ConstraintFk* fk : fkList)
+    for (SqlQueryModelColumn::ConstraintFk*& fk : fkList)
     {
         col = wrapObjIfNeeded(fk->foreignColumn);
         src = wrapObjIfNeeded(fk->foreignTable);
@@ -289,7 +258,7 @@ QString SqlQueryItemDelegate::getSqlForFkEditor(SqlQueryItem* item) const
             continue;
 
         srcCols = resolver.getTableColumns(src);
-        for (const QString& srcCol : srcCols)
+        for (QString& srcCol : srcCols)
         {
             if (fk->foreignColumn.compare(srcCol, Qt::CaseInsensitive) == 0)
                 continue; // Exclude matching column. We don't want the same column several times.
@@ -347,26 +316,16 @@ qlonglong SqlQueryItemDelegate::getRowCountForFkEditor(Db* db, const QString& qu
     return result->getSingleCell().toLongLong();
 }
 
-int SqlQueryItemDelegate::getFkViewHeaderWidth(SqlQueryView* fkView, bool includeScrollBar) const
-{
-    int wd = fkView->horizontalHeader()->length();
-    if (includeScrollBar && fkView->verticalScrollBar()->isVisible())
-        wd += fkView->verticalScrollBar()->width();
-
-    return wd;
-}
-
 QWidget* SqlQueryItemDelegate::getFkEditor(SqlQueryItem* item, QWidget* parent, const SqlQueryModel* model) const
 {
-    QString sql = getSqlForFkEditor(item);
-
     Db* db = model->getDb();
     bool countingError = false;
+    QString sql = FkComboBox::getSqlForFkEditor(db, item->getColumn(), item->getValue());
     qlonglong rowCount = getRowCountForFkEditor(db, sql, &countingError);
-    if (rowCount > MAX_ROWS_FOR_FK)
+    if (rowCount > FkComboBox::MAX_ROWS_FOR_FK)
     {
         notifyWarn(tr("Foreign key for column %2 has more than %1 possible values. It's too much to display in drop down list. You need to edit value manually.")
-                   .arg(MAX_ROWS_FOR_FK).arg(item->getColumn()->column));
+                       .arg(FkComboBox::MAX_ROWS_FOR_FK).arg(item->getColumn()->column));
 
         return getEditor(item->getValue().userType(), parent);
     }
@@ -377,151 +336,9 @@ QWidget* SqlQueryItemDelegate::getFkEditor(SqlQueryItem* item, QWidget* parent, 
         return nullptr;
     }
 
-    QComboBox *cb = new QComboBox(parent);
-    cb->setEditable(true);
-
-    // Prepare combo dropdown view.
-    SqlQueryView* comboView = new SqlQueryView();
-    comboView->setSimpleBrowserMode(true);
-    comboView->setMaximumWidth(QGuiApplication::primaryScreen()->size().width());
-
-    fkViewParentItemSize = model->getView()->horizontalHeader()->sectionSize(item->index().column());
-    connect(comboView->horizontalHeader(), &QHeaderView::sectionResized, [this, comboView, model](int, int, int)
-    {
-        if (!model->isAllDataLoaded())
-            return;
-
-        updateComboViewGeometry(comboView, false);
-    });
-
-    SqlQueryModel* comboModel = new SqlQueryModel(comboView);
-    comboModel->setView(comboView);
-
-    // Mapping of model to cb, so we can update combo when data arrives.
-    modelToFkCombo[comboModel] = cb;
-    connect(cb, &QComboBox::destroyed, [this, comboModel](QObject*)
-    {
-        modelToFkCombo.remove(comboModel);
-        modelToFkInitialValue.remove(comboModel);
-    });
-
-    // When execution is done, update combo.
-    connect(comboModel, SIGNAL(aboutToLoadResults()), this, SLOT(fkDataAboutToLoad()));
-    connect(comboModel, SIGNAL(executionSuccessful()), this, SLOT(fkDataReady()));
-    connect(comboModel, SIGNAL(executionFailed(QString)), this, SLOT(fkDataFailed(QString)));
-
-    // Setup combo, model, etc.
-    cb->setModel(comboModel);
-    cb->setView(comboView);
-    cb->setModelColumn(1);
-    cb->setCurrentText(item->getValue().toString());
-    cb->view()->viewport()->installEventFilter(new FkComboShowFilter(this, comboView, cb));
-    cb->view()->verticalScrollBar()->installEventFilter(new FkComboShowFilter(this, comboView, cb));
-    if (!item->getValue().isNull())
-        cb->lineEdit()->selectAll();
-
-    comboModel->setHardRowLimit(MAX_ROWS_FOR_FK);
-    comboModel->setCellDataLengthLimit(FK_CELL_LENGTH_LIMIT);
-    comboModel->setDb(db);
-    comboModel->setQuery(sql);
-    comboModel->setAsyncMode(true);
-    comboModel->executeQuery();
-
-    comboView->verticalHeader()->setVisible(false);
-    comboView->horizontalHeader()->setVisible(true);
-    comboView->setSelectionMode(QAbstractItemView::SingleSelection);
-    comboView->setSelectionBehavior(QAbstractItemView::SelectRows);
-
+    int dropDownViewMinWidth = model->getView()->horizontalHeader()->sectionSize(item->index().column());
+    FkComboBox* cb = new FkComboBox(parent, dropDownViewMinWidth);
+    cb->init(db, item->getColumn());
+    cb->setValue(item->getValue());
     return cb;
-}
-
-void SqlQueryItemDelegate::fkDataAboutToLoad()
-{
-    SqlQueryModel* model = dynamic_cast<SqlQueryModel*>(sender());
-    if (!modelToFkCombo.contains(model))
-        return;
-
-    QComboBox* cb = modelToFkCombo[model];
-    modelToFkInitialValue[model] = cb->currentText();
-}
-
-void SqlQueryItemDelegate::fkDataReady()
-{
-    SqlQueryModel* model = dynamic_cast<SqlQueryModel*>(sender());
-    if (!modelToFkCombo.contains(model))
-    {
-        qDebug() << "Deceived fkDataReady() call when there is no longer a combo linked to the model.";
-        return;
-    }
-
-    SqlQueryView* queryView = model->getView();
-
-    queryView->horizontalHeader()->setSectionHidden(0, true);
-    queryView->resizeColumnsToContents();
-    queryView->resizeRowsToContents();
-
-    updateComboViewGeometry(queryView, true);
-
-    // Set selected combo value to initial value from the cell
-    QComboBox* cb = modelToFkCombo[model];
-    QString valueFromQueryModel = modelToFkInitialValue[model].toString();
-
-    if (model->rowCount() > 0)
-    {
-        QModelIndex startIdx = model->index(0, cb->modelColumn());
-        QModelIndex endIdx = model->index(model->rowCount() - 1, cb->modelColumn());
-        QModelIndexList idxList = model->findIndexes(startIdx, endIdx, SqlQueryItem::DataRole::VALUE, valueFromQueryModel, 1, true);
-
-        if (idxList.size() > 0)
-        {
-            cb->setCurrentIndex(idxList.first().row());
-        }
-        else
-        {
-            cb->setCurrentIndex(-1);
-            cb->setEditText(valueFromQueryModel);
-        }
-    }
-    else
-        cb->setEditText(valueFromQueryModel);
-}
-
-void SqlQueryItemDelegate::fkDataFailed(const QString &errorText)
-{
-    notifyWarn(tr("Cannot edit this cell. Details: %1").arg(errorText));
-}
-
-void SqlQueryItemDelegate::updateComboViewGeometry(SqlQueryView* comboView, bool initial) const
-{
-    int wd = getFkViewHeaderWidth(comboView, true);
-    comboView->setMinimumWidth(qMin(qMax(fkViewParentItemSize, wd), comboView->maximumWidth()));
-
-    if (initial && wd < comboView->minimumWidth())
-    {
-        // First time, upon showing up
-        int currentSize = comboView->horizontalHeader()->sectionSize(1);
-        int gap = comboView->minimumWidth() - wd;
-        comboView->horizontalHeader()->resizeSection(1, currentSize + gap);
-    }
-
-    QWidget* container = comboView->parentWidget();
-    if (container->width() > comboView->minimumWidth())
-    {
-        container->setMaximumWidth(comboView->minimumWidth());
-        container->resize(comboView->minimumWidth(), container->height());
-    }
-}
-
-SqlQueryItemDelegate::FkComboShowFilter::FkComboShowFilter(const SqlQueryItemDelegate* delegate, SqlQueryView* comboView, QObject* parent)
-    : QObject(parent), delegate(delegate), comboView(comboView)
-{
-}
-
-bool SqlQueryItemDelegate::FkComboShowFilter::eventFilter(QObject* obj, QEvent* event)
-{
-    UNUSED(obj);
-    if (event->type() == QEvent::Show)
-        delegate->updateComboViewGeometry(comboView, true);
-
-    return false;
 }